Create funds-transfer.sh
The code is a bash script that automates the process of transferring funds from one phone number to another. The script first gets the cookie and token from the API, then it gets the curl commands for each step of the funds transfer flow, and finally it executes the curl commands.

The get_headers function gets the cookie and token from the API. The get_curl_command function gets the curl command for a given URL, cookie, and token. The execute_curl_commands function executes the curl commands.

The curl_commands variable contains a list of curl commands. Each curl command corresponds to a step in the funds transfer flow. For example, the first curl command sends the message *126# to the API. The second curl command selects option 1 from the menu. The third curl command selects option 1 from the submenu. And so on.

The execute_curl_commands function executes the curl commands one by one. The -s option tells curl to be silent, so that the script does not print the output of the curl commands to the console.

The code can be used to automate the process of transferring funds from a bash script. This can be useful for businesses that need to transfer funds to a large number of customers, or for individuals who want to automate their funds transfer transactions.

Here is a more detailed explanation of each function:

get_headers(): This function gets the cookie and token from the API. The API is called with the Accept header set to a specific value. The response from the API is then parsed to extract the cookie and token.
get_curl_command(): This function gets the curl command for a given URL, cookie, and token. The curl command is constructed with the Cookie header set to the cookie, the x-requested-with header set to XMLHttpRequest, and the __RequestVerificationToken header set to the token.
execute_curl_commands(): This function executes the curl commands. The curl commands are executed one by one, with the -s option set to silence the output of the curl commands.
The code for the funds transfer flow is as follows:

The script sends the message *126# to the API.
The API responds with a menu of options.
The script selects option 1 from the menu, which is the funds transfer option.
The API prompts the user for the recipient's phone number.
The script enters the recipient's phone number.
The API prompts the user for the amount to transfer.
The script enters the amount to transfer.
The API prompts the user for their PIN.
The script enters their PIN.
The API transfers the funds.