Create funds-transfer.py
Here is a step-by-step explanation of how the script works to complete a fund transfers after sim swaps :

The script gets the cookie and token from the API.
The script creates a list of curl commands.
The script executes the curl commands one by one.
The last curl command submits the fund transfer request.
The script can be used to automate the fund transfer process. This can be helpful for Hackers that need to make a large number of fund transfers after large numbers of sim swap.